Extracted TimeTo2PM to date
